feat: move weak entity  validation tests to dedicated file
